Merge 1.20 back into main

+++ Common/src/main/java/at/petrak/hexcasting/api/misc/MediaConstants.java
@@ -1,10 +1,10 @@
 package at.petrak.hexcasting.api.misc;
 
 public final class MediaConstants {
-    public static final int DUST_UNIT = 10000;
-    public static final int SHARD_UNIT = 5 * DUST_UNIT;
-    public static final int CRYSTAL_UNIT = 10 * DUST_UNIT;
+    public static final long DUST_UNIT = 10000;
+    public static final long SHARD_UNIT = 5 * DUST_UNIT;
+    public static final long CRYSTAL_UNIT = 10 * DUST_UNIT;
 
-    public static final int QUENCHED_SHARD_UNIT = 3 * CRYSTAL_UNIT;
-    public static final int QUENCHED_BLOCK_UNIT = 4 * QUENCHED_SHARD_UNIT;
+    public static final long QUENCHED_SHARD_UNIT = 3 * CRYSTAL_UNIT;
+    public static final long QUENCHED_BLOCK_UNIT = 4 * QUENCHED_SHARD_UNIT;
 }
